Jacat/Adjust/requirements.txt
prefect
prefect-gcp
google-cloud-storage
pathlib
google-cloud-bigquery
python-dotenv


